[update] implementando o FiledDAteTime com passagem de padrao por parametro no construtor
git-svn-id: https://svn.autogeral.com.br/sped/EFDPisCofins@155 56544dd1-a790-4711-ae29-638a3112a704

+++ test/br/com/jcomputacao/sped/efd/pisCofins/Randomize.java
@@ -44,7 +44,7 @@ public class Randomize {
         String Str = "";
         while (Str.length() < tamanho) {
             Random r = new Random();
-            Str +=  r.nextDouble();
+            Str +=  r.nextLong();
         }    
       return Double.parseDouble(Str);
     }
